dreamirc: python-modernize: lib2to3.fixes.fix_long

+++ dreamirc/src/protocols/irc.py
@@ -1511,7 +1511,7 @@ def dccParseAddress(address):
         pass
     else:
         try:
-            address = long(address)
+            address = int(address)
         except ValueError:
             raise IRCBadMessage,\
                   "Indecipherable address %r" % (address,)
@@ -1737,7 +1737,7 @@ def dccDescribe(data):
         pass
     else:
         try:
-            address = long(address)
+            address = int(address)
         except ValueError:
             pass
         else:
